update: renamed stringUtils file and replaced typography with div
