+++ packages/ui/src/utils/index.ts
@@ -14,3 +14,4 @@ export * from './getApproveAsMultiTx'
 export * from './getAsMultiTx'
 export * from './getErrorMessageReservedFunds'
 export * from './extinsicErrorChecks'
+export * from './camelcasetoString'
